@@ -14,15 +14,15 @@
     "@testing-library/user-event": "^12.1.10",
     "@types/jest": "^26.0.15",
     "@types/node": "^12.0.0",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
+    "@types/react": "18",
+    "@types/react-dom": "18",
     "@types/react-router-dom": "^5.1.8",
     "autoprefixer": "^9",
     "axios": "^0.21.1",
     "immer": "^9.0.6",
     "postcss": "^7",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "18",
+    "react-dom": "18",
     "react-query": "^3.19.1",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
@@ -71,6 +71,9 @@
       "last 1 safari version"
     ]
   },
+  "resolutions": {
+    "@types/react": "17.0.30"
+  },
   "devDependencies": {
     "@commitlint/cli": "^13.1.0",
     "@commitlint/config-conventional": "^13.1.0",
@@ -93,5 +96,6 @@
     "prettier": "^2.3.2",
     "pretty-quick": "^3.1.1",
     "ts-jest": "^26.5.5"
-  }
+  },
+  "packageManager": "yarn@3.2.1"
 }
